nudge

+++ programs/bpf/c/src/multiple_static/multiple_static.c
@@ -3,7 +3,7 @@
 static const char msg[] = "This is a message";
 static const char msg2[] = "This is a different message";
 
-extern uint32_t entrypoint(const uint8_t *input) {
+extern uint64_t entrypoint(const uint8_t *input) {
   sol_log((char*)msg);
   sol_log((char*)msg2);
   return SUCCESS;
