+++ programs/bpf/c/src/struct_ret/struct_ret.c
@@ -15,7 +15,7 @@ static struct test_struct __attribute__ ((noinline)) test_function(void) {
   return s;
 }
 
-extern uint32_t entrypoint(const uint8_t* input) {
+extern uint64_t entrypoint(const uint8_t* input) {
   struct test_struct s = test_function();
   sol_log("foobar");
   if (s.x + s.y + s.z == 12 ) {
