+++ programs/bpf/rust/alloc/src/lib.rs
@@ -3,11 +3,11 @@
 #[macro_use]
 extern crate alloc;
 extern crate solana_sdk;
-use solana_sdk::{entrypoint::SUCCESS, info};
+use solana_sdk::{info, program_error::SUCCESS};
 use std::{alloc::Layout, mem};
 
 #[no_mangle]
-pub extern "C" fn entrypoint(_input: *mut u8) -> u32 {
+pub extern "C" fn entrypoint(_input: *mut u8) -> u64 {
     unsafe {
         // Confirm large allocation fails
